requirements.txt
paramiko
paho-mqtt
